requirements.txt
tavily-python==0.7.6
python-dotenv==1.1.0
langchain==0.3.25
langchain-openai==0.3.23
langchain-anthropic==0.3.11
langchain-core==0.3.65
langgraph==0.4.8
pydantic==2.11.7
requests==2.32.3
typing-extensions==4.12.2
flask==3.1.0
flask-cors==5.0.0
setuptools==70.0.0
python-jose
starlette>=0.40.0
langgraph-prebuilt==0.2.2
langchain-tavily==0.2.6